+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Effort in sidebar
 - Compendiums: SWNR Classes, SWNR Spells, MTA Spells
 - Add badges to spell list to show rote if or praxis
+- Filter MTA spells by arcanum by clicking the name in the magic tab sidebar
 
 ## [1.6.1] - 2024-04-10